requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-formtools==2.2
django-heroku==0.3.1
django-static-fontawesome==5.10.2
django-taggit==1.2.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
